requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
pyspark==3.4.1
pandas==2.0.0
numpy==1.24.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
scikit-learn==1.3.2
streamlit>=1.32.0
plotly>=5.20.0
